tmcl_host_interface
Added root _send and _recv functions to tmcl_host_interface.
With this and the recent changes in PyTrinamic a new tmcl_interface can
inherit the tmcl_interface for functionality to communicate with TMCL
slave and tmcl_host_interface for functionality to communicate with TMCL
host. Both require the _send and _recv functions to be implemented.